Revert "feat(lazer): improve protocol types and modules (#2934)"
This reverts commit 1101ad2d86a1baf9e5273cbf83632210117d1700.
