+++ lazer/sdk/rust/protocol/src/message.rs
@@ -1,6 +1,6 @@
 use {
     self::format_magics_le::{EVM_FORMAT_MAGIC, SOLANA_FORMAT_MAGIC},
-    crate::api::ParsedPayload,
+    crate::router::ParsedPayload,
     anyhow::{bail, Context},
     byteorder::{ByteOrder, ReadBytesExt, WriteBytesExt, BE, LE},
     derive_more::From,
