+++ lazer/sdk/rust/protocol/src/payload.rs
@@ -1,10 +1,12 @@
-use crate::{
-    price::Price,
-    rate::Rate,
-    time::{DurationUs, TimestampUs},
-    ChannelId, PriceFeedId, PriceFeedProperty,
-};
+//! Types representing binary encoding of signable payloads and signature envelopes.
+
+use crate::time::DurationUs;
 use {
+    super::router::{PriceFeedId, PriceFeedProperty},
+    crate::{
+        router::{ChannelId, Price, Rate},
+        time::TimestampUs,
+    },
     anyhow::bail,
     byteorder::{ByteOrder, ReadBytesExt, WriteBytesExt, BE, LE},
     serde::{Deserialize, Serialize},
@@ -229,12 +231,12 @@ fn write_option_price<BO: ByteOrder>(
     mut writer: impl Write,
     value: Option<Price>,
 ) -> std::io::Result<()> {
-    writer.write_i64::<BO>(value.map_or(0, |v| v.mantissa_i64()))
+    writer.write_i64::<BO>(value.map_or(0, |v| v.0.get()))
 }
 
 fn read_option_price<BO: ByteOrder>(mut reader: impl Read) -> std::io::Result<Option<Price>> {
     let value = NonZeroI64::new(reader.read_i64::<BO>()?);
-    Ok(value.map(Price::from_nonzero_mantissa))
+    Ok(value.map(Price))
 }
 
 fn write_option_rate<BO: ByteOrder>(
@@ -244,7 +246,7 @@ fn write_option_rate<BO: ByteOrder>(
     match value {
         Some(value) => {
             writer.write_u8(1)?;
-            writer.write_i64::<BO>(value.mantissa())
+            writer.write_i64::<BO>(value.0)
         }
         None => {
             writer.write_u8(0)?;
@@ -256,7 +258,7 @@ fn write_option_rate<BO: ByteOrder>(
 fn read_option_rate<BO: ByteOrder>(mut reader: impl Read) -> std::io::Result<Option<Rate>> {
     let present = reader.read_u8()? != 0;
     if present {
-        Ok(Some(Rate::from_mantissa(reader.read_i64::<BO>()?)))
+        Ok(Some(Rate(reader.read_i64::<BO>()?)))
     } else {
         Ok(None)
     }
