+++ lazer/sdk/rust/protocol/src/serde_str.rs
@@ -1,6 +1,6 @@
 pub mod option_price {
     use {
-        crate::price::Price,
+        crate::router::Price,
         serde::{de::Error, Deserialize, Deserializer, Serialize, Serializer},
         std::num::NonZeroI64,
     };
@@ -10,7 +10,7 @@ pub mod option_price {
         S: Serializer,
     {
         value
-            .map(|price| price.mantissa_i64().to_string())
+            .map(|price| price.0.get().to_string())
             .serialize(serializer)
     }
 
@@ -22,7 +22,7 @@ pub mod option_price {
         if let Some(value) = value {
             let value: i64 = value.parse().map_err(D::Error::custom)?;
             let value = NonZeroI64::new(value).ok_or_else(|| D::Error::custom("zero price"))?;
-            Ok(Some(Price::from_nonzero_mantissa(value)))
+            Ok(Some(Price(value)))
         } else {
             Ok(None)
         }
